+++ CHANGELOG.md
@@ -33,6 +33,7 @@ This file follows the best practices from [keepachangelog.com](http://keepachang
 ### Security
 
 - Updated Paperclip gem [#1836](https://github.com/sharetribe/sharetribe/pull/1836)
+- Unauthorized users were able to upload new listing images [#1866](https://github.com/sharetribe/sharetribe/pull/1866)
 
 ## [5.6.0] - 2016-03-11